pyproject.toml
[project]
name = "mvp-ds-rag-hw"
version = "0.1.0"
description = ""
authors = [
    {name = "Rinat",email = "oblom.ru@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu",
    "httpx (>=0.28.1,<0.29.0)",
    "PyMuPDF",
    "numpy<2",
    "sentence-transformers (>=2.2,<4)",
    "torch",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
